refactor(seat): 좌석 선점 트랜잭션-락 경계 불일치 해결 및 안정성 개선 [GRGB-263] (#146)
* refactor(seat): SeatHoldService 트랜잭션 로직을 별도 Bean으로 분리하여 락-커밋 순서 보장

기존에는 SeatHoldService.createOrRefreshHold()에 @Transactional이 걸려 있어 finally 블록에서 unlock이 먼저 실행되고, Spring AOP 프록시가 이후에 커밋하는 구조였음.
이로 인해 다른 스레드가 커밋되지 않은 데이터를 읽을 수 있는 위험이 존재.

- SeatHoldTransactionalService를 별도 빈으로 분리하여 트랜잭션 커밋이 완료된 후 finally에서 unlock되도록 순서를 보장함.
- 추가로 releaseUserActiveHolds에서 delete 후 flush()를 호출하여 JPA flush 순서(INSERT → UPDATE → DELETE)로 인한 unique constraint violation을 방지함

* refactor(seat): SeatAssignmentService 트랜잭션 로직을 별도 빈으로 분리

SeatHoldService와 동일한 트랜잭션-락 경계 문제 해결.

기존 doAssignAndHold()는 같은 클래스 내부 호출(self-invocation)이라 @Transactional AOP 프록시가 적용되지 않는 문제도 함께 존재했음.

- SeatAssignmentTransactionalService를 별도 빈으로 분리하여 트랜잭션이 정상적으로 동작하고, 커밋 완료 후 락이 해제되도록 보장함.
- cleanupExistingHolds에서도 delete 후 flush()를 추가하여 JPA flush 순서 문제를 방지함.

* fix(seat): SeatHoldLockManager WAIT_TIME 100ms에서 500ms로 상향

100ms는 피크 트래픽 시 Redis 네트워크 지연이나 DB 트랜잭션 처리 시간이 겹치면 정상 요청도 락 획득에 실패할 수 있는 값임.
- 다른 스레드의 트랜잭션 처리 시간: 락을 잡은 스레드가 DB 조회 → 검증 → 상태 변경 → 커밋까지 하는데 보통 50~100ms 정도 소요. GC pause나 DB 부하가 겹치면 100ms 초과 가능

500ms로 상향하여 사용자 체감에 영향 없이 불필요한 실패를 방지함.
- 좌석 선점 트랜잭션이 정상적으로 처리되는 시간(~100ms)의 약 5배 여유
- 사용자 체감상 500ms 대기는 거의 인지 불가 (HTTP 요청 전체 응답이 ~600ms)
- 그러면서도 진짜 데드락이나 비정상 상황에서는 빠르게 실패

* test(seat): 트랜잭션-락 분리 구조에 맞춰 테스트 재작성

- SeatHoldServic/SeatAssignmentService는 validation, 락 획득/해제, 위임 호출만 검증하도록 변경.
- SeatHoldTransactionalService/SeatAssignmentTransactionalService에 비즈니스 로직 테스트(만료 연장, 좌석 교체, 충돌 예외 등)를 분리하여 작성.
- 예외 발생 시에도 락이 반드시 해제되는지 검증하는 테스트 추가.

* docs(seat): 일반좌석 hold service 코드에 JavaDoc 문서화 추가
